feat: landing page refactor for agenda page
