+++ src/app/agenda/agenda.component.html
@@ -1,0 +1,1 @@
+<p>agenda works!</p>
