+++ src/app/app-routing.module.ts
@@ -1,7 +1,15 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [];
+import { AgendaComponent } from '../app/agenda/agenda.component';
+import { HomeComponent } from '../app/home/home.component';
+
+const routes: Routes = [
+  { path: 'agenda', component: AgendaComponent },
+  { path: 'home', component: HomeComponent },
+  { path: '', redirectTo: '/home', pathMatch: 'full' },
+  { path: '**', redirectTo: '/home', pathMatch: 'full' }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
