+++ src/app/app.component.html
@@ -1,2 +1,1 @@
-<app-mobile *ngIf="showMobile" [conferenceTime]="conferenceTime" [timezone]="timezone"></app-mobile>
-<app-desktop *ngIf="!showMobile" [conferenceTime]="conferenceTime" [timezone]="timezone"></app-desktop>
+<router-outlet></router-outlet>
