+++ src/app/desktop/desktop.component.html
@@ -3,7 +3,7 @@
         <img src="../../assets/FC_LOGO.png" alt="Logo">
         <span class="title">Fundamental<br>Conference</span>
         <div class="links">
-            <a href="">
+            <a routerLink="/agenda">
                 <span class="top-label">Agenda</span><br>
                 <span class="hidden-bottom-label">Agenda</span>
             </a>
