+++ src/app/home/home.component.html
@@ -1,0 +1,2 @@
+<app-mobile *ngIf="showMobile" [conferenceTime]="conferenceTime" [timezone]="timezone"></app-mobile>
+<app-desktop *ngIf="!showMobile" [conferenceTime]="conferenceTime" [timezone]="timezone"></app-desktop>
